more work on sdl backend

+++ src/aether/core/input.h
@@ -4,7 +4,9 @@
 
 #ifdef AETHER_USE_ALLEGRO
 #include "allegro/al_keycode.h"
-#elif AETHER_USE_VITA
+#elif AETHER_USE_SDL
+#include "sdl/sdl_keycode.h"
+#elif AETHER_USE_3DS
 #include "3ds/3ds_keycode.h"
 #endif
 
@@ -12,6 +14,8 @@
 
 #ifdef AETHER_USE_ALLEGRO
 #include "allegro/al_keycode.h"
+#elif AETHER_USE_SDL
+#include "sdl/sdl_keycode.h"
 #elif AETHER_USE_3DS
 #include "3ds/3ds_keycode.h"
 #endif
@@ -22,7 +26,7 @@ namespace core {
 struct MouseState {
     int x;
     int y;
-    uint8_t buttons;
+    uint32_t buttons;
 };
 
 void enable_grab_mouse();
